+++ tests/unit/test_hooks.py
@@ -17,7 +17,7 @@ from tests import unittest
 from functools import partial
 
 from botocore.hooks import HierarchicalEmitter, first_non_none_response
-from botocore.hooks import AliasedEventEmitter
+from botocore.hooks import EventAliaser
 
 
 class TestHierarchicalEventEmitter(unittest.TestCase):
@@ -62,16 +62,20 @@ class TestHierarchicalEventEmitter(unittest.TestCase):
         self.assertEqual(calls, ['foo.bar.baz', 'foo.bar', 'foo'])
 
 
-class TestAliasedEventEmitter(unittest.TestCase):
+class TestAliasedEmitter(unittest.TestCase):
     def setUp(self):
         self.hook_calls = []
 
     def hook(self, **kwargs):
         self.hook_calls.append(kwargs)
 
+    def get_emitter(self, event_aliases):
+        emitter = HierarchicalEmitter()
+        return EventAliaser(emitter, event_aliases)
+
     def test_event_emitted(self):
         aliases = {'bar': 'bear'}
-        emitter = AliasedEventEmitter(event_aliases=aliases)
+        emitter = self.get_emitter(event_aliases=aliases)
         emitter.register('foo.bear.baz', self.hook)
         emitter.emit('foo.bear.baz')
         calls = [e['event_name'] for e in self.hook_calls]
@@ -79,23 +83,39 @@ class TestAliasedEventEmitter(unittest.TestCase):
 
     def test_aliased_event_emitted(self):
         aliases = {'bar': 'bear'}
-        emitter = AliasedEventEmitter(event_aliases=aliases)
+        emitter = self.get_emitter(event_aliases=aliases)
         emitter.register('foo.bear.baz', self.hook)
         emitter.emit('foo.bar.baz')
         calls = [e['event_name'] for e in self.hook_calls]
         self.assertEqual(calls, ['foo.bear.baz'])
 
+    def test_alias_with_dots_emitted(self):
+        aliases = {'api.bar': 'bear'}
+        emitter = self.get_emitter(event_aliases=aliases)
+        emitter.register('foo.bear.baz', self.hook)
+        emitter.emit('foo.api.bar.baz')
+        calls = [e['event_name'] for e in self.hook_calls]
+        self.assertEqual(calls, ['foo.bear.baz'])
+
     def test_aliased_event_registered(self):
         aliases = {'bar': 'bear'}
-        emitter = AliasedEventEmitter(event_aliases=aliases)
+        emitter = self.get_emitter(event_aliases=aliases)
         emitter.register('foo.bar.baz', self.hook)
+        emitter.emit('foo.bear.baz')
+        calls = [e['event_name'] for e in self.hook_calls]
+        self.assertEqual(calls, ['foo.bear.baz'])
+
+    def test_aliased_event_with_dots_registered(self):
+        aliases = {'api.bar': 'bear'}
+        emitter = self.get_emitter(event_aliases=aliases)
+        emitter.register('foo.api.bar.baz', self.hook)
         emitter.emit('foo.bear.baz')
         calls = [e['event_name'] for e in self.hook_calls]
         self.assertEqual(calls, ['foo.bear.baz'])
 
     def test_event_unregistered(self):
         aliases = {'bar': 'bear'}
-        emitter = AliasedEventEmitter(event_aliases=aliases)
+        emitter = self.get_emitter(event_aliases=aliases)
 
         emitter.register('foo.bar.baz', self.hook)
         emitter.emit('foo.bear.baz')
@@ -110,7 +130,7 @@ class TestAliasedEventEmitter(unittest.TestCase):
 
     def test_aliased_event_unregistered(self):
         aliases = {'bar': 'bear'}
-        emitter = AliasedEventEmitter(event_aliases=aliases)
+        emitter = self.get_emitter(event_aliases=aliases)
 
         emitter.register('foo.bar.baz', self.hook)
         emitter.emit('foo.bear.baz')
@@ -119,6 +139,21 @@ class TestAliasedEventEmitter(unittest.TestCase):
 
         self.hook_calls = []
         emitter.unregister('foo.bar.baz', self.hook)
+        emitter.emit('foo.bear.baz')
+        calls = [e['event_name'] for e in self.hook_calls]
+        self.assertEqual(calls, [])
+
+    def test_aliased_event_with_dots_unregistered(self):
+        aliases = {'api.bar': 'bear'}
+        emitter = self.get_emitter(event_aliases=aliases)
+
+        emitter.register('foo.api.bar.baz', self.hook)
+        emitter.emit('foo.bear.baz')
+        calls = [e['event_name'] for e in self.hook_calls]
+        self.assertEqual(calls, ['foo.bear.baz'])
+
+        self.hook_calls = []
+        emitter.unregister('foo.api.bar.baz', self.hook)
         emitter.emit('foo.bear.baz')
         calls = [e['event_name'] for e in self.hook_calls]
         self.assertEqual(calls, [])